\name{print.rTableICC}
\alias{print.rTableICC}

\title{
Print Summary of Data Generation Process and Generated Data
}
\description{
Prints summary information on data generation process and generated data.
}
\usage{
\method{print}{rTableICC}(x,...)
}

\arguments{
  \item{x}{
an object including information and random table to be printed.
}
  \item{...}{
other arguments.
}
}
\author{
Haydar Demirhan

Maintainer: Haydar Demirhan <haydar.demirhan@rmit.edu.au>
}

